Update: Top Up

+++ src/foodordersystem/Model/Dwallet.java
@@ -3,7 +3,7 @@ package foodordersystem.Model;
 import foodordersystem.Model.User;
 
 public class Dwallet extends User {
-    public Dwallet (int id, String username, int credit) {
+    public Dwallet (int id, String username, double credit) {
         super(id, username, credit);
     }
 }
